requirements.txt
tensorflow
tensorflow_intel==2.11.0
numpy==1.24.1
nltk==3.8.1
sqlite3==3.37.2
markdown==3.4.2
google-generativeai==0.2.2
Django==3.2.12
django-widget-tweaks==1.4.11
